refactor(continue-watching): replace dedicated tab with a Library row
Drop the standalone Continue tab and surface in-progress items as a
horizontal "Continue Watching" shelf at the top of the gallery (Library)
root, hidden inside folders and self-hiding when empty. Reuses the
existing local watch-progress store (getRecentProgress + fetchItemsByIds)
and the VideoGridItem progress bar; adds a cardWidth prop so cards size
to a fixed width in the row. Removes getItemLayout so the variable-height
list header doesn't break scroll math.
